fix: fix race condition in suback code
After client sent a subscribe packet, it waits for suback from the socket. But __listen also reads from the same socket on a separate thread waiting for any publish packets. This caused a race condition where the suback was sometimes accepted by __listen and never got to the main thread.